test(dedupe): consolidate telemetry_streaming tests (5 → 4 files)
Consolidate telemetry_streaming test files in tests/unit/gpt_trader/features/live_trade/engines:
- test_telemetry_streaming_emit_and_enable.py + task_completion.py → test_telemetry_streaming.py
- test_telemetry_streaming_run_loop.py → test_telemetry_streaming_loop.py (rename)
- test_telemetry_streaming_background_restart.py → test_telemetry_streaming_restart.py (rename)
- test_telemetry_streaming_schedule_and_start_stop.py → test_telemetry_streaming_lifecycle.py (rename)

Cluster: 7ae3ae24a668
Files: 5 → 4 (-1)
Tests: 43 passing

Note: 240-line limit prevented further consolidation.
