+++ tests/unit/gpt_trader/features/live_trade/engines/test_telemetry_streaming_lifecycle.py
