+++ tests/unit/gpt_trader/features/live_trade/engines/test_telemetry_streaming_restart.py
